+++ init.el
@@ -41,9 +41,6 @@ This function should only modify configuration layer settings."
      ;;
      ;; Layers added in alphabetic order
 
-     ;; Enable asciidoc layer for editing asciidoc content e.g. docs.cider.mx editing
-     asciidoc
-
      ;; Add tool tips to show doc string of functions
      ;; Show snippets in the auto-completion popup
      ;; Show suggestions by most commonly used
@@ -51,26 +48,24 @@ This function should only modify configuration layer settings."
                       auto-completion-enable-help-tooltip t
                       auto-completion-enable-snippets-in-popup t
                       auto-completion-enable-sort-by-usage t
-                      ;; auto-completion-idle-delay 0.0
+                      auto-completion-idle-delay 0.0
+                      auto-completion-minimum-prefix-length 1
                       ;; auto-completion-complete-with-key-sequence "fd"
                       )
 
-     ;; To have auto-completion on as soon as you start typing
-     ;; (auto-completion :variables auto-completion-idle-delay nil)
-
      ;; https://develop.spacemacs.org/layers/+lang/clojure/README.html
      (clojure :variables
-              ;; clojure-backend 'cider               ;; use cider and disable lsp
-              ;; clojure-enable-linters 'clj-kondo    ;; clj-kondo included in lsp
-              clojure-enable-kaocha-runner t          ;; enable Kaocha test runner
-              cider-repl-display-help-banner nil      ;; disable help banner
-              cider-print-fn 'puget                   ;; pretty printing with sorted keys / set values
+              ;; clojure-backend 'cider                 ; use cider and disable lsp
+              clojure-enable-kaocha-runner t            ; enable Kaocha test runner
+              cider-repl-display-help-banner nil        ; disable help banner
+              cider-print-fn 'puget                     ; pretty printing with sorted keys / set values
               clojure-indent-style 'align-arguments
               clojure-align-forms-automatically t
-              clojure-toplevel-inside-comment-form t  ;; evaluate expressions in comment as top level
-              cider-result-overlay-position 'at-point ;; results shown right after expression
+              clojure-toplevel-inside-comment-form t ; clashes with LSP
+              cider-result-overlay-position 'at-point   ; results shown right after expression
               cider-overlays-use-font-lock t
-              cider-repl-buffer-size-limit 100        ;; limit lines shown in REPL buffer
+              cider-repl-buffer-size-limit 100          ; limit lines shown in REPL buffer
+              nrepl-use-ssh-fallback-for-remote-hosts t ; connect via ssh to remote hosts
               )
 
      (go :variables
@@ -102,20 +97,26 @@ This function should only modify configuration layer settings."
 
      ;; SPC g s opens Magit git client full screen (q restores previous layout)
      ;; show word-granularity differences in current diff hunk
+     ;; delta-plugin requires git-detlta install
+     ;; https://develop.spacemacs.org/layers/+source-control/git/README.html#magit-plugins
      (git :variables
           git-magit-status-fullscreen t
           magit-diff-refine-hunk t
-          git-enable-magit-todos-plugin t)
+          git-enable-magit-todos-plugin t
+
+          ;; Install https://github.com/dandavison/delta for language syntax highlights in diff
+          ;; git-enable-magit-delta-plugin t
+          )
 
      ;; graphviz - open-source graph declaration system
      ;; Used to generated graphs of Clojure project dependencies
      ;; https://develop.spacemacs.org/layers/+lang/graphviz/README.html
      graphviz
 
-     ;; helm-follow-mode sticky - remembers use of C-c C-f
-     ;; - follow mode previews when scrolling through a helm list
-     (helm :variables
-           helm-follow-mode-persistent t)
+     ;; add variable `helm-follow-mode-persistent t' for helm-follow-mode sticky
+     ;; helm follow mode previews when scrolling through a helm list
+     ;; - remembers use of C-c C-f in helm pop-up
+     helm
 
      html
      javascript
@@ -124,14 +125,20 @@ This function should only modify configuration layer settings."
      ;; Language server protocol with minimal visual impact
      ;; https://practical.li/spacemacs/install-spacemacs/clojure-lsp/
      (lsp :variables
-          lsp-ui-doc-enable nil       ;; disable all doc popups
-          lsp-ui-sideline-enable nil  ;; disable sideline bar for less distraction
-          treemacs-space-between-root-nodes nil  ;; no spacing in treemacs views
-          lsp-log-io t  ; Log client-server json communication
+          lsp-headerline-breadcrumb-enable t              ; Breadcrumb trail
+          lsp-headerline-breadcrumb-segments '(symbols)   ; namespace & symbols, no file path
+          lsp-ui-peek-enable t                            ; popups for refs, errors, symbols, etc.
+          lsp-semantic-tokens-enable t                    ; enhance syntax highlight
+          lsp-treemacs-error-list-current-project-only t  ; limit errors to current project
+          lsp-idle-delay 0.2                              ; smooth LSP features response
+          lsp-eldoc-enable-hover nil                      ; disable all hover actions
+          lsp-ui-doc-enable nil                           ; doc hover popups
+          lsp-ui-sideline-enable nil                      ; sidebar code actions visual indicator
+          treemacs-space-between-root-nodes nil           ; spacing in treemacs views
+          lsp-log-io t                                    ; Log client-server json communication
           )
 
-     (markdown :variables
-               markdown-live-preview-engine 'vmd)
+     markdown
 
      ;; Editing multiple lines of text concurrently
      ;; `g r' menu in Emacs normal state
@@ -141,7 +148,7 @@ This function should only modify configuration layer settings."
      (org :variables
           org-enable-github-support t
           org-enable-bootstrap-support t
-          org-enable-reveal-js-support t
+          ;; org-enable-reveal-js-support t
           org-want-todo-bindings t
           org-enable-org-journal-support t
           org-journal-dir "~/projects/journal/"
@@ -163,7 +170,8 @@ This function should only modify configuration layer settings."
 
      ;; SPC ' runs shell in a popup buffer
      (shell :variables
-            shell-default-shell 'eshell
+            shell-default-shell 'vterm
+            shell-default-term-shell "/usr/bin/zsh"
             shell-default-height 30
             shell-default-position 'bottom)
 
@@ -177,11 +185,16 @@ This function should only modify configuration layer settings."
      ;; Configuration: https://github.com/seagle0128/doom-modeline#customize
      (spacemacs-modeline :variables
                          doom-modeline-height 12
-                         doom-modeline-major-mode-color-icon t
+                         doom-modeline-bar-width 0                  ; size of bar icon at start of modeline
+                         doom-modeline-buffer-state-icon t          ; disk icon for unsaved changes (default)
+                         doom-modeline-major-mode-color-icon t      ; color icon of major mode
                          doom-modeline-buffer-file-name-style 'relative-to-project
-                         doom-modeline-display-default-persp-name t
-                         doom-modeline-minor-modes nil
-                         doom-modeline-modal-icon nil)
+                         doom-modeline-display-default-persp-name t ; layout name
+                         doom-modeline-minor-modes nil              ; show minor modes
+                         doom-modeline-modal nil                    ; show Evil state icon/text
+                         ;; doom-modeline-modal-icon t              ; icon when t, ascii when nil
+                         doom-modeline-buffer-encoding nil          ; remove UTF-8, etc.
+                         )
 
      ;; Spell as you type with Flyspell package,
      ;; requires external command - ispell, hunspell, aspell
@@ -199,11 +212,6 @@ This function should only modify configuration layer settings."
                treemacs-use-filewatch-mode t
                treemacs-use-follow-mode t)
 
-     ;; Customise the Spacemacs themes
-     ;; https://develop.spacemacs.org/layers/+themes/theming/README.html
-     ;; Code in dotspacemacs/user-init to reduce size of modeline
-     theming
-
      ;; Support font ligatures (fancy symbols) in all modes
      ;; 'prog-mode for only programming languages
      ;; including text-mode may cause issues with org-mode and magit
@@ -220,7 +228,7 @@ This function should only modify configuration layer settings."
      (yaml :variables
            yaml-enable-lsp t)
 
-     ) ;; End of dotspacemacs-configuration-layers
+     ) ; End of dotspacemacs-configuration-layers
 
 
    ;; List of additional packages that will be installed without being wrapped
@@ -231,7 +239,7 @@ This function should only modify configuration layer settings."
    ;; `dotspacemacs/user-config'. To use a local version of a package, use the
    ;; `:location' property: '(your-package :location "~/path/to/your-package/")
    ;; Also include the dependencies as they will not be resolved automatically.
-   dotspacemacs-additional-packages '(keycast elcord)
+   dotspacemacs-additional-packages '(clojure-essential-ref)
 
    ;; A list of packages that cannot be updated.
    dotspacemacs-frozen-packages '()
@@ -365,9 +373,8 @@ It should only modify the values of Spacemacs settings."
    ;; pair of numbers, e.g. `(recents-by-project . (7 .  5))', where the first
    ;; number is the project limit and the second the limit on the recent files
    ;; within a project.
-   dotspacemacs-startup-lists '((projects . 3)
-                                (todos . 5)
-                                (bookmarks . 20))
+   dotspacemacs-startup-lists '((recents . 5)
+                                (projects . 7))
 
    ;; True if the home buffer should respond to resize events. (default t)
    dotspacemacs-startup-buffer-responsive t
@@ -386,10 +393,10 @@ It should only modify the values of Spacemacs settings."
    ;; Default major mode for a new empty buffer. Possible values are mode
    ;; names such as `text-mode'; and `nil' to use Fundamental mode.
    ;; (default `text-mode')
-   dotspacemacs-new-empty-buffer-major-mode 'org-mode
+   dotspacemacs-new-empty-buffer-major-mode 'text-mode
 
    ;; Default major mode of the scratch buffer (default `text-mode')
-   dotspacemacs-scratch-mode 'org-mode
+   dotspacemacs-scratch-mode 'text-mode
 
    ;; If non-nil, *scratch* buffer will be persistent. Things you write down in
    ;; *scratch* buffer will be saved and restored automatically.
@@ -406,14 +413,10 @@ It should only modify the values of Spacemacs settings."
    ;; List of themes, the first of the list is loaded when spacemacs starts.
    ;; Press `SPC T n' to cycle to the next theme in the list (works great
    ;; with 2 themes variants, one dark and one light)
-
-   dotspacemacs-themes '(kaolin-valley-dark
+   dotspacemacs-themes '(doom-gruvbox
                          doom-gruvbox-light
-                         doom-solarized-light
-                         doom-sourcerer
-                         doom-solarized-dark
-                         spacemacs-light
-                         spacemacs-dark)
+                         spacemacs-dark
+                         spacemacs-light)
 
    ;; Set the theme for the Spaceline. Supported themes are `spacemacs',
    ;; `all-the-icons', `custom', `doom', `vim-powerline' and `vanilla'. The
@@ -469,11 +472,11 @@ It should only modify the values of Spacemacs settings."
    dotspacemacs-distinguish-gui-tab nil
 
    ;; Name of the default layout (default "Default")
-   dotspacemacs-default-layout-name "Global"
+   dotspacemacs-default-layout-name "Default"
 
    ;; If non-nil the default layout name is displayed in the mode-line.
    ;; (default nil)
-   dotspacemacs-display-default-layout t
+   dotspacemacs-display-default-layout nil
 
    ;; If non-nil then the last auto saved layouts are resumed automatically upon
    ;; start. (default nil)
@@ -545,7 +548,7 @@ It should only modify the values of Spacemacs settings."
    ;; A value from the range (0..100), in increasing opacity, which describes
    ;; the transparency level of a frame when it's active or selected.
    ;; Transparency can be toggled through `toggle-transparency'. (default 90)
-   dotspacemacs-active-transparency 100
+   dotspacemacs-active-transparency 90
 
    ;; A value from the range (0..100), in increasing opacity, which describes
    ;; the transparency level of a frame when it's inactive or deselected.
@@ -601,10 +604,9 @@ It should only modify the values of Spacemacs settings."
                                                    pdf-view-mode
                                :size-limit-kb 1000)
 
-
    ;; Code folding method. Possible values are `evil', `origami' and `vimish'.
    ;; (default 'evil)
-   dotspacemacs-folding-method 'vimish
+   dotspacemacs-folding-method 'evil
 
    ;; If non-nil and `dotspacemacs-activate-smartparens-mode' is also non-nil,
    ;; `smartparens-strict-mode' will be enabled in programming modes.
@@ -627,7 +629,7 @@ It should only modify the values of Spacemacs settings."
 
    ;; If non-nil, start an Emacs server if one is not already running.
    ;; (default nil)
-   dotspacemacs-enable-server t
+   dotspacemacs-enable-server nil
 
    ;; Set the emacs server socket location.
    ;; If nil, uses whatever the Emacs default is, otherwise a directory path
@@ -638,7 +640,7 @@ It should only modify the values of Spacemacs settings."
 
    ;; If non-nil, advise quit functions to keep server open when quitting.
    ;; (default nil)
-   dotspacemacs-persistent-server t
+   dotspacemacs-persistent-server nil
 
    ;; List of search tool executable names. Spacemacs uses the first installed
    ;; tool of the list. Supported tools are `rg', `ag', `pt', `ack' and `grep'.
@@ -665,7 +667,7 @@ It should only modify the values of Spacemacs settings."
    ;; performance issues, instead of calculating the frame title by
    ;; `spacemacs/title-prepare' all the time.
    ;; (default "%I@%S")
-   dotspacemacs-frame-title-format nil
+   dotspacemacs-frame-title-format "%I@%S"
 
    ;; Format specification for setting the icon title format
    ;; (default nil - same as frame-title-format)
@@ -674,7 +676,7 @@ It should only modify the values of Spacemacs settings."
    ;; Color highlight trailing whitespace in all prog-mode and text-mode derived
    ;; modes such as c++-mode, python-mode, emacs-lisp, html-mode, rst-mode etc.
    ;; (default t)
-   dotspacemacs-show-trailing-whitespace nil
+   dotspacemacs-show-trailing-whitespace t
 
    ;; Delete whitespace while saving buffer. Possible values are `all'
    ;; to aggressively delete empty line and long sequences of whitespace,
@@ -707,7 +709,7 @@ It should only modify the values of Spacemacs settings."
    ;; Run `spacemacs/prettify-org-buffer' when
    ;; visiting README.org files of Spacemacs.
    ;; (default nil)
-   dotspacemacs-pretty-docs t
+   dotspacemacs-pretty-docs nil
 
    ;; If nil the home buffer shows the full path of agenda items
    ;; and todos. If non-nil only the file name is shown.
@@ -722,7 +724,8 @@ This function defines the environment variables for your Emacs session. By
 default it calls `spacemacs/load-spacemacs-env' which loads the environment
 variables declared in `~/.spacemacs.env' or `~/.spacemacs.d/.spacemacs.env'.
 See the header of this file for more information."
-  (spacemacs/load-spacemacs-env))
+  (spacemacs/load-spacemacs-env)
+)
 
 (defun dotspacemacs/user-init ()
   "Initialization for user code:
@@ -735,31 +738,15 @@ If you are unsure, try setting them in `dotspacemacs/user-config' first."
   ;; simplifying version control of the Spacemacs configuration file
   (setq custom-file (file-truename (concat dotspacemacs-directory "emacs-custom-settings.el")))
   (load custom-file)
+)
 
-
-  ;; custom theme modification
-  ;; spacemacs - overriding default height of modeline
-  ;; doom-gruvbox - subtle lsp symbol highlight
-  (setq-default
-    theming-modifications
-      '((spacemacs-light
-          (mode-line :height 0.92)
-          (mode-line-inactive :height 0.92))
-        (doom-solarized-light
-         (mode-line :height 0.92)
-         (mode-line-inactive :height 0.92))
-        (doom-gruvbox-light
-         (lsp-face-highlight-read :background nil :weight bold)
-         (command-log-command :foreground "firebrick")
-         (command-log-key :foreground "dark magenta"))))
-
-  )  ;; End of dotspacemacs/user-int
 
 (defun dotspacemacs/user-load ()
   "Library to load while dumping.
 This function is called only while dumping Spacemacs configuration. You can
 `require' or `load' the libraries of your choice that will be included in the
-dump.")
+dump."
+)
 
 
 (defun dotspacemacs/user-config ()
@@ -769,623 +756,35 @@ configuration.
 Put your configuration code here, except for variables that should be set
 before packages are loaded."
 
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Keeping Helm history clean
-  (setq history-delete-duplicates t)
-  (setq extended-command-history
-        (delq nil (delete-dups extended-command-history)))
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Connecting to a reomote nREPL server
-  (setq nrepl-use-ssh-fallback-for-remote-hosts t)
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Keycast - show Emacs commands in mode line
-  (use-package keycast
-                :commands keycast-mode
-                :config
-                (define-minor-mode keycast-mode
-                  "Show current command and its key binding in the mode line."
-                  :global t
-                  (if keycast-mode
-                      (progn
-                        (add-hook 'pre-command-hook 'keycast-mode-line-update t)
-                        (add-to-list 'mode-line-misc-info '("" mode-line-keycast "    "))
-                        )
-                    (remove-hook 'pre-command-hook 'keycast-mode-line-update)
-                    (setq global-mode-string (remove '("" mode-line-keycast " ") mode-line-misc-info))))
-                )
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Emacs text rendering optimizations
-  ;; https://200ok.ch/posts/2020-09-29_comprehensive_guide_on_handling_long_lines_in_emacs.html
-
-  ;; Only render text left to right
-  (setq-default bidi-paragraph-direction 'left-to-right)
-
-  ;; Disable Bidirectional Parentheses Algorithm
-  (if (version<= "27.1" emacs-version)
-      (setq bidi-inhibit-bpa t))
-
-  ;; Files with known long lines
-  ;; SPC f l to open files literally to disable most text processing
-
-  ;; So long mode when Emacs thinks a file would affect performance
-  (if (version<= "27.1" emacs-version)
-      (global-so-long-mode 1))
-
-  ;; End of: Emacs text rendering optimizations
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Doom theme settings
-  ;; (setq doom-gruvbox-light-variant "hard")
-  ;;
-  (defun practicalli/setup-custom-doom-modeline ()
-    (doom-modeline-set-modeline 'practicalli-modeline 'default))
-  ;;
-  (with-eval-after-load 'doom-modeline
-    (doom-modeline-def-modeline 'practicalli-modeline
-      '(workspace-name window-number modals persp-name buffer-info matches remote-host vcs)
-      '(misc-info repl lsp buffer-position))
-    (practicalli/setup-custom-doom-modeline))
-
-  ;; checker = flycheck results (not working)
-  ;; buffer-position
-  ;; word-count - number of words in current buffer
-  ;; parrot
-  ;; selection-info
-  ;; repl - shows status of Cloure repl (not working)
-  ;; process ??
-  ;; debug
-  ;; misc-info  - used for keycast
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; User key bindings
-  ;;
-  ;; org-journal - create a new journal entry - `, j' in org-journal mode
-  (spacemacs/set-leader-keys "oj" 'org-journal-new-entry)
-  ;;
-  ;; Toggle workspaces forward/backwards
-  (spacemacs/set-leader-keys "ow" 'eyebrowse-next-window-config)
-  (spacemacs/set-leader-keys "oW" 'eyebrowse-last-window-config)
-
-  ;; Revert buffer - loads in .dir-locals.el changes
-  (spacemacs/set-leader-keys "oR" 'revert-buffer)
-  ;;
-  ;; Keycast mode - show key bindings and commands in mode line
-  (spacemacs/set-leader-keys "ok" 'keycast-mode)
-
-  ;; Replace Emacs Tabs key bindings with Workspace key bindings
-  (with-eval-after-load 'evil-maps
-    (when (featurep 'tab-bar)
-      (define-key evil-normal-state-map "gt" nil)
-      (define-key evil-normal-state-map "gT" nil)))
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Over-ride Spacemacs defaults
-  ;;
-  ;; Set new location for file bookmarks, SPC f b
-  ;; Default: SPACEMACSDIR.cache/bookmarks
-  ;; (setq bookmark-default-file (concat dotspacemacs-directory "bookmarks"))
-  ;;
-  ;; Set new location for recent save files
-  ;; Default: SPACEMACSDIR.cache/recentf
-  ;; (setq recentf-save-file (concat dotspacemacs-directory "recentf") )
-  ;;
-
-  ;; native line numbers taking up lots of space?
-  (setq-default display-line-numbers-width nil)
-  ;;
-  ;; replace / search with helm-swoop in Evil normal state
-  (evil-global-set-key 'normal "/" 'helm-swoop)
-  ;;
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Magit - forge configuration
-  ;;
-  ;; Set the files that are searched for writing tokens
-  ;; by default ~/.authinfo will be used
-  ;; and write a token in unencrypted format
-  ;; (setq auth-sources '("~/.authinfo.gpg"))
-
-  ;; Use XDG_CONFIG_HOME location or HOME
-  (setq auth-sources (list
-                      (concat (getenv "XDG_CONFIG_HOME") "/authinfo.gpg")
-                      "~/.authinfo.gpg"))
-
-  ;;
-  ;; Configure number of topics show, open and closed
-  ;; use negative number to toggle the view of closed topics
-  ;; using `SPC SPC forge-toggle-closed-visibility'
-  (setq  forge-topic-list-limit '(100 . -10))
-  ;; set closed to 0 to never show closed issues
-  ;; (setq  forge-topic-list-limit '(100 . 0))
-  ;;
-  ;; GitHub user and organization accounts owned
-  ;; used by @ c f  to create a fork
-  (setq forge-owned-accounts
-        '(("practicalli" "practicalli-john"
-           "ClojureBridgeLondon" "ldnclj"
-           "clojure-hacks"
-           "reclojure")))
-  ;; To blacklist specific accounts,
-  ;; over-riding forge-owned-accounts
-  ;; (setq forge-owned-blacklist
-  ;;       '(("bad-hacks" "really-bad-hacks")))
-  ;;
-  ;; End of Magit - forge configuration
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Safe structural editing
-  ;; for all major modes
-  (spacemacs/toggle-evil-safe-lisp-structural-editing-on-register-hooks)
-  ;; for clojure layer only (comment out line above)
-  ;; (spacemacs/toggle-evil-safe-lisp-structural-editing-on-register-hook-clojure-mode)
-  ;;
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Version Control configuration - Git, etc
-  ;;
-  ;; diff-hl - diff hightlights in right gutter as you type
-  (diff-hl-flydiff-mode)
-  ;;
-  ;; Load in magithub features after magit package has loaded
-  ;; (use-package magithub
-  ;;   :after magit
-  ;;   :config (magithub-feature-autoinject t))
-  ;;
-  ;; Use Spacemacs as the $EDITOR (or $GIT_EDITOR) for git commits messages
-  ;; when using git commit on the command line
-  ;; (global-git-commit-mode t)
-  ;;
-  ;; Set locations of all your Git repositories
-  ;; with a number to define how many sub-directories to search
-  ;; `SPC g L' - list all Git repositories in the defined paths,
-  (setq magit-repository-directories
-        '(("~/projects/" . 2)))
-  ;;
-  ;; end of version control configuration
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Org-mode configuration
-  ;;
-  ;; I should write a toggle function to show descriptive or literate links in Org-mode
-  ;;(setq org-descriptive-links nil)
-  ;;
-  ;; Org-reveal - define were reveal.js files can be found
-  ;; (I place reveal.js files in same directory as I write the org files)
-  (setq org-reveal-root "")
-  ;;
-  ;; Define the location of the file to hold tasks
-  (with-eval-after-load 'org
-    (setq org-default-notes-file "~/Dropbox/todo-list.org"))
-  ;;
-  ;; Define a kanban style set of stages for todo tasks
-  (with-eval-after-load 'org
-    (setq org-todo-keywords
-         '((sequence "TODO" "DOING" "BLOCKED" "REVIEW" "|" "DONE" "ARCHIVED"))))
-  ;;
-  ;; The default keywords all use the same colour.
-  ;; Make the states easier to distinguish by using different colours
-  ;; Using X11 colour names from: https://en.wikipedia.org/wiki/Web_colors
-  ;; Setting colours (faces) using the `org-todo-keyword-faces' defcustom function
-  ;; https://github.com/tkf/org-mode/blob/master/lisp/org-faces.el#L376
-  ;; Using `with-eval-after-load' as a hook to call this setting when org-mode is run
-  ;;
-  (with-eval-after-load 'org
-    (setq org-todo-keyword-faces
-          '(("TODO" . "SlateGray")
-            ("DOING" . "DarkOrchid")
-            ("BLOCKED" . "Firebrick")
-            ("REVIEW" . "Teal")
-            ("DONE" . "ForestGreen")
-            ("ARCHIVED" .  "SlateBlue"))))
-  ;;
-  ;;
-  ;; Set TODO keyword faces if over-ridden by theme.
-  (defun practicalli/set-todo-keyword-faces ()
-    (interactive)
-    (setq hl-todo-keyword-faces
-          '(("TODO" . "SlateGray")
-            ("DOING" . "DarkOrchid")
-            ("BLOCKED" . "Firebrick")
-            ("REVIEW" . "Teal")
-            ("DONE" . "ForestGreen")
-            ("ARCHIVED" .  "SlateBlue"))))
-  ;;
-  ;;
-  ;; Progress Logging
-  ;; When a TODO item enters DONE, add a CLOSED: property with current date-time stamp
-  (with-eval-after-load 'org
-    (setq org-log-done 'time))
-  ;;
-  ;;
-  ;; customize org-mode's checkboxes with unicode symbols
-  (add-hook
-   'org-mode-hook
-   (lambda ()
-     "Beautify Org Checkbox Symbol"
-     (push '("[ ]" . "☐") prettify-symbols-alist)
-     (push '("[X]" . "☑" ) prettify-symbols-alist)
-     (push '("[-]" . "❍" ) prettify-symbols-alist)
-     (prettify-symbols-mode)))
-  ;;
-  ;; Markdown mode hook for orgtbl-mode minor mode
-  ;; (add-hook 'markdown-mode-hook 'turn-on-orgtbl)
-  ;;
-  ;; Turn on visual-line-mode for Org-mode only
-  ;; (add-hook 'org-mode-hook 'turn-on-visual-line-mode)
-  ;;
-  ;; use org-re-reveal instead of org-reveal (which hasnt been updated in ages and breaks org-mode 9.2)
-  ;; (use-package org-re-reveal :after org)
-  ;;
-  ;; End of Org-mode Configuration
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Clojure configurations
-  ;;
-  ;; Do not indent single ; comment characters
-  (add-hook 'clojure-mode-hook (lambda () (setq-local comment-column 0)))
-
-  ;; Auto-indent code automatically
-  ;; https://emacsredux.com/blog/2016/02/07/auto-indent-your-code-with-aggressive-indent-mode/
-  ;; (add-hook 'clojure-mode-hook #'aggressive-indent-mode)
-
-  ;; Lookup functions in Clojure - The Essentail Reference book
-  ;; https://github.com/p3r7/clojure-essential-ref
-  (spacemacs/set-leader-keys "oh" 'clojure-essential-ref)
-  ;;
-  ;;
-  ;; toggle reader macro sexp comment
-  ;; toggles the #_ characters at the start of an expression
-  (defun clojure-toggle-reader-comment-sexp ()
-    (interactive)
-    (let* ((point-pos1 (point)))
-      (evil-insert-line 0)
-      (let* ((point-pos2 (point))
-             (cmtstr "#_")
-             (cmtstr-len (length cmtstr))
-             (line-start (buffer-substring-no-properties point-pos2 (+ point-pos2 cmtstr-len)))
-             (point-movement (if (string= cmtstr line-start) -2 2))
-             (ending-point-pos (+ point-pos1 point-movement 1)))
-        (if (string= cmtstr line-start)
-            (delete-char cmtstr-len)
-          (insert cmtstr))
-        (goto-char ending-point-pos)))
-    (evil-normal-state))
-  ;;
-  ;; Assign keybinding to the toggle-reader-comment-sexp function
-  (define-key global-map (kbd "C-#") 'clojure-toggle-reader-comment-sexp)
-  ;;
-  ;;
-  ;; Toggle view of a clojure `(comment ,,,) block'
-  (defun clojure-hack/toggle-comment-block (arg)
-    "Close all top level (comment) forms. With universal arg, open all."
-    (interactive "P")
-    (save-excursion
-      (goto-char (point-min))
-      (while (search-forward-regexp "^(comment\\>" nil 'noerror)
-        (call-interactively
-         (if arg 'evil-open-fold
-           'evil-close-fold)))))
-  ;;
-  (evil-define-key 'normal clojure-mode-map
-    "zC" 'clojure-hack/toggle-comment-block
-    "zO" (lambda () (interactive) (clojure-hack/toggle-comment-block 'open)))
-  ;;
-  ;; end of clojure configuration
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Web-mode configuration
-  ;;
-  ;; Changing auto indent size for languages in html layer (web mode) to 2 (defaults to 4)
-  (defun web-mode-indent-2-hook ()
-    "Indent settings for languages in Web mode, markup=html, css=css, code=javascript/php/etc."
-    (setq web-mode-markup-indent-offset 2)
-    (setq web-mode-css-indent-offset  2)
-    (setq web-mode-code-indent-offset 2))
-  ;;
-  (add-hook 'web-mode-hook  'web-mode-indent-2-hook)
-  ;;
-  ;; End of Web-mode configuration
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Eshell visual enhancements
-  ;;
-  ;; Add git status visual labels
-  ;;
-  (require 'dash)
-  (require 's)
-  ;;
-  (defmacro with-face (STR &rest PROPS)
-    "Return STR propertized with PROPS."
-    `(propertize ,STR 'face (list ,@PROPS)))
-  ;;
-  (defmacro esh-section (NAME ICON FORM &rest PROPS)
-    "Build eshell section NAME with ICON prepended to evaled FORM with PROPS."
-    `(setq ,NAME
-           (lambda () (when ,FORM
-                        (-> ,ICON
-                            (concat esh-section-delim ,FORM)
-                            (with-face ,@PROPS))))))
-  ;;
-  (defun esh-acc (acc x)
-    "Accumulator for evaluating and concatenating esh-sections."
-    (--if-let (funcall x)
-        (if (s-blank? acc)
-            it
-          (concat acc esh-sep it))
-      acc))
-  ;;
-  (defun esh-prompt-func ()
-    "Build `eshell-prompt-function'"
-    (concat esh-header
-            (-reduce-from 'esh-acc "" eshell-funcs)
-            "\n"
-            eshell-prompt-string))
-  ;;
-  ;;
-  ;; Unicode icons on Emacs
-  ;; `list-character-sets' and select unicode-bmp
-  ;; scroll through bitmaps list to find the one you want
-  ;; some bitmaps seem to change
-  ;;
-  (esh-section esh-dir
-               "\xf07c"  ;  (faicon folder)
-               (abbreviate-file-name (eshell/pwd))
-               '(:foreground "olive" :bold bold :underline t))
-  ;;
-  (esh-section esh-git
-               "\xf397"  ;  (git branch icon)
-               (magit-get-current-branch)
-               '(:foreground "maroon"))
-  ;;
-  ;; (esh-section esh-python
-  ;;              "\xe928"  ;  (python icon)
-  ;;              pyvenv-virtual-env-name)
-  ;;
-  (esh-section esh-clock
-               ""  ;  (clock icon)
-               (format-time-string "%H:%M" (current-time))
-               '(:foreground "forest green"))
-  ;;
-  ;; Below I implement a "prompt number" section
-  (setq esh-prompt-num 0)
-  (add-hook 'eshell-exit-hook (lambda () (setq esh-prompt-num 0)))
-  (advice-add 'eshell-send-input :before
-              (lambda (&rest args) (setq esh-prompt-num (incf esh-prompt-num))))
-  ;;
-  ;;
-  ;; "\xf0c9"  ;  (list icon)
-  (esh-section esh-num
-               "\x2130"  ;  ℰ (eshell icon)
-               (number-to-string esh-prompt-num)
-               '(:foreground "brown"))
-  ;;
-  ;; Separator between esh-sections
-  (setq esh-sep " ")  ; or " | "
-  ;;
-  ;; Separator between an esh-section icon and form
-  (setq esh-section-delim "")
-  ;;
-  ;; Eshell prompt header
-  (setq esh-header "\n ")  ; or "\n┌─"
-  ;;
-  ;; Eshell prompt regexp and string. Unless you are varying the prompt by eg.
-  ;; your login, these can be the same.
-  (setq eshell-prompt-regexp " \x2130 ")   ; or "└─> "
-  (setq eshell-prompt-string " \x2130 ")   ; or "└─> "
-  ;;
-  ;; Choose which eshell-funcs to enable
-  ;; (setq eshell-funcs (list esh-dir esh-git esh-python esh-clock esh-num))
-  ;; (setq eshell-funcs (list esh-dir esh-git esh-clock esh-num))
-  (setq eshell-funcs (list esh-dir esh-git))
-
-  ;; Enable the new eshell prompt
-  (setq eshell-prompt-function 'esh-prompt-func)
-
-  ;; End of Eshell
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Shell configuration
-  ;;
-  ;; Use zsh for default multi-term shell
-  ;; (setq multi-term-program "/usr/bin/zsh")
-  ;;
-  ;; End of Shell configuration
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; MacOSX
-  ;; Disable touchpad zoom gestures
-  ;;
-  ;; (define-key global-map (kbd "<magnify-up>") nil)
-  ;; (define-key global-map (kbd "<magnify-down>") nil)
-  ;;
-  ;; (defun practicalli-nothing ()
-  ;;   (interactive)
-  ;;   (message "Buttons are not toys") )
-  ;;
-  ;; (define-key global-map (kbd "<magnify-up>") 'practicalli-nothing)
-  ;;
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Spaceline Doom theme settings
-  ;; https://seagle0128.github.io/doom-modeline/
-  ;; Configuration set in layer variables
-  ;;
-  ;; Set height of the modeline - will resize to height of text
-  ;; (setq doom-modeline-height 12)
-
-  ;; The left hand bar in the modeline
-  ;; setting to zero shows a large box outline
-  ;; (setq doom-modeline-bar-width 1)
-
-  ;; Determine style of current filename / path displayed
-  ;; default: auto
-  ;; (setq doom-modeline-buffer-file-name-style 'relative-to-project)
-
-  ;; default perspective name displayed in the mode-line.
-  ;; (setq doom-modeline-display-default-persp-name t)
-
-  ;; Do not show buffer encoding
-  ;; (setq doom-modeline-buffer-encoding nil)
-
-  ;; display GitHub notifications (requires `ghub' package)
-  ;; (setq doom-modeline-github t)
-  ;; The interval of checking GitHub.
-  ;; (setq doom-modeline-github-interval (* 30 60))
-
-  ;; GNUs notifications - default t
-  ;; (setq doom-modeline-gnus nil)
-
-  ;; IRC notifications - default t
-  ;; (setq doom-modeline-irc nil)
-
-  ;; Environment versions - default t
-  ;; (setq doom-modeline-env-version t)
-
-  ;; Use ascii rather than icon for modal state (more specific)
-  ;; Icon not changing for doom-solarized-light theme
-  ;; - icon changes color for doom-gruvbox-light theme
-  ;; (setq doom-modeline-modal-icon nil)
-  ;; End of Spaceline Doom theme settings
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Configuration no longer used
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Workarounds and bug fixes - temporary hopefully
-  ;;
-  ;; From a community question - not advisable
-  ;; Set a different key binding for evil lisp state
-  ;; (with-eval-after-load 'evil-lisp-state
-  ;;   (spacemacs/set-leader-keys "k" evil-lisp-state-map))
-  ;;
-  ;; evil-escape - switch between insert and normal state
-  ;; (setq-default evil-escape-key-sequence "fd")
-  ;;
-  ;; Undo history size limit, triggering garbage collection
-  ;; Updating all defaults by a power of 10 (adding another zero at the end)
-  ;; default in spacemacs is 80000
-  ;; (setq undo-limit 400000)
-  ;;
-  ;; default in spacemacs is 120000
-  ;; (setq undo-strong-limit 6000000)
-  ;;
-  ;; default in spacemacs is 12000000
-  ;; (setq undo-strong-limit 60000000)
-  ;;
-  ;;
-  ;; disable undo-tree as it seems to be loosing history
-  ;; (global-undo-tree-mode -1)
-  ;;
-  ;; TODO: try explicitly saving history
-  ;; (setq undo-tree-auto-save-history t)
-  ;;
-  ;; TODO: try setting undo-tree tmp files location
-  ;; (setq undo-tree-history-directory-alist '(("." . "~/var/emacs/undo")))
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Literal Searching Configuration
-  ;;
-  ;; Literal search, rather than regex, in spacemacs search - helm-ag
-  ;; (setq-default helm-grep-ag-command-option "-Q")
-  ;;
-  ;; End of Searching Configuration
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Systemd user service
-  ;;
-  ;; Use the exec-path-from-shell package to get PATH, MANPATH
-  ;; and the environment variables from your zsh or bash rc-files.
-  ;;
-  ;; (setq exec-path-from-shell-variables
-  ;;       (append exec-path-from-shell-variables
-  ;;               (list "TERM"
-  ;;                     "RUST_SRC_PATH"
-  ;;                     "…"
-  ;;                     )))
-  ;; (exec-path-from-shell-initialize)
-  ;;
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; Neotree configuration
-  ;;
-  ;; Display neotree on the right rather than left (default)
-  ;; (setq neo-window-position 'right)
-  ;;
-  ;; End of Neotree configuration
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-  ;; old-school emacs style keybindings that i am replacing with nicer spacemacs alternatives
-  ;;
-  ;; jr0cket: text scaling keybindings - use spc z x =/-
-  ;; (define-key global-map (kbd "c-+") 'text-scale-increase)
-  ;; (define-key global-map (kbd "c--") 'text-scale-decrease)
-  ;;
-  ;; smartparens keybindings - use lisp-state, spc k menu instead
-  ;; (define-key global-map (kbd "c-)") 'sp-forward-slurp-sexp)
-  ;; (define-key global-map (kbd "c-(") 'sp-backward-slurp-sexp)
-  ;; (define-key global-map (kbd "m-)") 'sp-forward-barf-sexp)
-  ;; (define-key global-map (kbd "m-(") 'sp-backward-barf-sexp)
-  ;;
-  ;; jr0cket: keybindings for cycling buffers
-  ;; use spc b n and spc b n instead
-  ;; (global-set-key [c-prior] 'previous-buffer)
-  ;; (global-set-key [c-next] 'next-buffer)
-  ;;
-  ;; jr0cket: remap multiple cursors to a pattern that is easier to remember
-  ;; learn iedit mode instead (its fantastic)
-  ;; (define-key global-map (kbd "c-c m c") 'mc/edit-lines)
-  ;;
-  ;; end of old-school bindings
-  ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
-  (setq insert-directory-program "gls" dired-use-ls-dired t)
-  (setq dired-listing-switches "-al --group-directories-first")
-
-  (require 'elcord)
-
-  )   ;; End of dot-spacemacs/user-config
+  ;; dotspacemacs/user-config divided into files
+  ;; - comment files to skip loading specific configuration
+
+  ;; General configuration
+  (setq user-config-file (file-truename (concat dotspacemacs-directory "user-config.el")))
+  (load user-config-file)
+
+  ;; Clojure Layer additionalb configuration
+  (setq clojure-config-file (file-truename (concat dotspacemacs-directory "clojure-config.el")))
+  (load clojure-config-file)
+
+  ;; Customise Theme and Mode-line
+  (setq theme-config-file (file-truename (concat dotspacemacs-directory "theme-config.el")))
+  (load theme-config-file)
+
+  ;; Version Control - Magit & Forge
+  (setq version-control-config-file (file-truename (concat dotspacemacs-directory "version-control-config.el")))
+  (load version-control-config-file)
+
+  ;; Org-mode customisation - notes and tasks
+  ;; (setq org-config-file (file-truename (concat dotspacemacs-directory "org-config.el")))
+  ;; (load org-config-file)
+
+  ;; EShell Customisation
+  ;; NOTE: Practicalli uses vterm for shell by default
+  ;; (setq eshell-config-file (file-truename (concat dotspacemacs-directory "eshell-config.el")))
+  ;; (load eshell-config-file)
+
+)
 
 
 ;; Do not write anything past this comment. This is where Emacs will